Backen-SaborBot-main/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
boto3>=1.28.57
botocore>=1.31.57
python-multipart==0.0.6
pydantic==1.10.11
sqlalchemy
pymysql
python-dotenv
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
